refactor(local-asr): extract presentational sub-components into LocalAsr/components.tsx — behavior unchanged
Pure move of 4 props-driven sub-components (FoundryPrepareProgressBlock,
DownloadProgressBlock, ModelRow, TestResultBlock) + ModelRowProps into
LocalAsr/components.tsx, and the shared RemoteSize type into LocalAsr/types.ts.
Imported back into index.tsx unchanged; no rendered-output or behavior change.
index.tsx 3497→2921 lines. tsc --noEmit / npm run build: green (bundle sizes identical).
