+++ openless-all/app/src/pages/LocalAsr/types.ts
@@ -1,0 +1,8 @@
+// Shared types for the Local ASR page and its sub-components.
+
+export interface RemoteSize {
+    totalBytes: number
+    fileCount: number
+    loading: boolean
+    error: string | null
+}
